this comit only adds some beauty to the NeoHooke material...
!!!HOWEVER!!!
Thomas and I have found a very, very... bad bug in the way the
pow-function (from math.h) operates on our cluster. It seems
that material and thus element evaluation time is slowed down
extremely in certain cases. Just to give you an idea: in terms
of element evaluation times we are talking about factors(!!!)
of 30-50 here. Currently we have only observed this catastrophic
behavior for the NeoHooke material, however the pow-function is
used at so many places in BACI that it is quite likely that
similar problems might occur elsewhere, too. 

It took us almost all afternoon to analyze the problem, and
still we are not completely sure what is the actual reason...
The best thing to do is probably to replace the corresponding
libraries (where the math stuff is contained) on our cluster
and try again if things are getting better! (Tobias, maybe
you can help us with that next week?)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15416 060e22d8-9d1c-0410-868a-e3a950d63d95