partially revised my yesterday's commit due to problems with Boost on kaiser. On kaiser a Boost 1.41 is installed in a non-standard way. With the yesterday's version of FindBoost.cmake the library paths got screwed up.
The behaviour now is:
1) use Boost from system (i.e. Boost 1.41 on kaiser, Boost 1.29 on hop&malt and Boost 1.34 on some workstations (FC8))
2) if no Boost is found on the system use the official LNM fallback Boost 1.46.1 from /lnm/lib/*/

There is a good reason to link Boost dynamically, since Boost is usually optimized for a specific system/hardware platform. Therefore it makes sense
to use a system-wide installed Boost before the LNM Boost from the network. However with Boost most of our machines are binary not compatible, that 
is most probably you cannot run baci that is compiled on your machine on another workstation. Note, that even if you use the same operating system (e.g. FC8)
it is not guaranteed that your BACI is binary compatible to other FC8 machines in our institute, since some of use have a local Boost and others not.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15447 060e22d8-9d1c-0410-868a-e3a950d63d95