Got rid of errors due to numerical zeros in phi-values and according bugs...
No phi-values are modified, but every check of the interface side of a point or node due to it's phi-value is (hopefully) done now by the "plusDomain"-function implemented in "xfem_utils.H" (which is the old "xfem_enums.H").

Adopted several files according to this change and fixed some bugs occuring due to "numerical positive or negative zeros".

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15457 060e22d8-9d1c-0410-868a-e3a950d63d95