Frame for topology optimization: - new problem type "topopt" - new material "opti_dens"
Just the rudimentary frame for optimization:
In "drt_opti" the routines are (or will be) called, mainly in the "opti_algorithm". Currently the first optimization step is called and contains of a call of the standard fluid with an additional reactive term. This term can be interpreted as an inverse porosity alpha which is computed due to the optimization density field. This field describes the percentage of fluid at every point of the domain and is thus a function rho(x) in [0,1] for every x. It will be optimized.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15462 060e22d8-9d1c-0410-868a-e3a950d63d95