some additions to "material" yoghurt:
For the time being, any simulation using yoghurt will need to be run as problem type "Low_Mach_Number_Flow", although the temperature represents a passive scalar in this case, as mentioned in the preceding commit message related to revision No. 15438. A test case for this material is available. However, I refrained from including this test case, since the present material does not represent a primarily-used material, on the one hand, and our test case series is already somewhat "stretched", on the other hand. Nevertheless, this test case might be immediately included if required.

(As aforementioned, this is based on developments by Georg Schlick in the context of his semester thesis.)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@15501 060e22d8-9d1c-0410-868a-e3a950d63d95