Functions (or their components) defined in input files can use the time "t" in the same way as a space dimension, e.g.
FUNCT1 EXPR 0 0.0 0 FUNCTION t*t-x*t

Actually two lines had to be changed. The time-dependent functions have already been read from input files and a time was set when evaluating a function. But somehow this time was changed to zero when going to the final evaluate function somewhere deep in the code...

Lots of function-evaluate routines were modified so that the correct time is used for evaluation.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16217 060e22d8-9d1c-0410-868a-e3a950d63d95