Commit fixes the problem, when multiple elasthyper materials with different length of NUMMAT are called, like
MAT 1 MAT_ElastHyper NUMMAT 1 MATIDS 10 DENS 1.0
MAT 2 MAT_ElastHyper NUMMAT 2 MATIDS 10 11 DENS 1.0

The problem was that the length of the IntVectorMaterialComponent was updated only once after the initialisation. By extending the query
if(length_ == -1) to
if( length_ == -1 or lengthname_ == 'NUMMAT')

the length of the IntVector for the second material will be updated.

A.N.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16398 060e22d8-9d1c-0410-868a-e3a950d63d95