Added the new Euler-Bernoulli-like beam element beam3ebtor. In contrary to the beam3eb element this element can also display torsion and is therefore a complete, geometrically exact Euler-Bernoulli beam element for initially (i.e. in the stress-free configuration) straight beams.
To be able to use this element correctly structural dynamic parameters need to be set to: LOADLIN Yes

For this element a special type of Neumann conditions, namely DESIGN POINT MOMENT EB CONDITIONS, are needed.

Each node of the element has 7 dofs: the nodal position vector (3 dofs), the nodal tangent vector (3 dofs) and the nodal twist angle (1 dof).

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16454 060e22d8-9d1c-0410-868a-e3a950d63d95