Fixed the problem if a simulation calculates with materials, which only partly include fibers.
If one processor calculates only materials without fibers the function visNames of io.cpp returns
{"Owner":1}

If one processor calculates materials with fibers the function visNames of io.cpp returns
{"Owner":1, "Fibers1":3}

Obvious this structure leads to errors in the postprocessing.

The solution was now to communicate parallel the merge of the gathered maps.

This is done by implementing the function GatherAll, which will be later called in io.cpp, in linalg_utils.h as a template
for template<typename T, typename U> std::map<T,U>
This function already existed for sets and vectors and was expanded for maps

Header of this function has the form:

template<typename T, typename U> void GatherAll(std::map<T,U>&      data,
                                       const        Epetra_Comm& comm)


For the implementation several functions in drt_parobject, drt_export had to be added.

I tested the behavior for a structural simulation with several processor and this testing delivered the same result. 
Since there are no test cases for post processing please tell me if you notice a strange behavior regarding your output.


Andi

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16470 060e22d8-9d1c-0410-868a-e3a950d63d95