re-established the usability of the good old and very helpful TRAP_FE flag.
When the TRAP_FE is actived in the configure file (default is off), 
a baci simulation stops immediately when a floating point exception occurs.
This may be a division by zero, an invalid floating point operation 
or the appearance of NaN or INF. Under- and overflow of doubles could be checked
as well.

Since several nightly test cases reveal problems, it might be
beneficial for all of us to have the TRAP_FE flag activated for our
regular testing.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16663 060e22d8-9d1c-0410-868a-e3a950d63d95