MORE SAFETY FOR BACI USERS:
duplicate definitions of the same input parameter are now detected 
and baci will stop with a dserror. It seems in case of duplicate 
parameter definitions in the input file 
always the last one was taken in the simulation.

I cleaned all affected input files for you!
(do not worry, I did not find a really bad bug).
No. 1 of removed duplicates is AZCONV found in various solver parameter lists.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16706 060e22d8-9d1c-0410-868a-e3a950d63d95