relocation of the recently introduced check for negative material parameters from Integrate() to CalcNewParameters(); sorry for that but its way better now
simplification of CheckOptStep() since it was done unnecessarily complicated due to the above
  
the drop tolerance given in the datfile now is meaningful for the residual as well as the gradient based
update; previously it was the error in the residual (without consideration of the update scheme used)
which was taken as stopping criterion for the optimization loop.

The latter modification has the great advantage, that for the gradient based scheme the error theoretically tends
to zero and the drop tolerance is not dependent of the problem as it is for the residual based scheme. In practice
this is influenced by the perturbation parameters and the general capability of the material to be fitted to represent
the physical behavior

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16754 060e22d8-9d1c-0410-868a-e3a950d63d95