Coupling of reduced-dimensional airways and 3D tissue models ============================================================
This is the first step towards an overall lung model consisting of reduced-D airways
and a continuum model for the parenchyma. The following partitioned solution scheme
is realized: the pressure at each outlet of the airway tree is used as a Neumann
boundary condition (orthopressure with the value being set from outside) in the
associated tissue region. After having solved the nonlinear solid problem, the air
flux is calculated based on the volume change of the tissue region and used as an
outflow condition of the associated airway problem.

First results look promising, but thorough testing is yet to come...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16756 060e22d8-9d1c-0410-868a-e3a950d63d95