bugfix no. 2 in cut_mesh.cpp
Uli's preselection of possible cuts between elements and sides was based on checking for overlaps between bounding boxes of the octTree's nodes with bounding boxes of the cut side. In case of to fine subdivisions of the octTree's bounding boxes, such a box can be fully contained in a background element and does not store any nodes adjacent to elements. Missing cut-pairs between elements and sides are the result. Sometimes the cut-algorithm did not crash but produced wrong volumecells and wrong inside-outside position etc.

We replaced the more efficient but defective preselection by a simple but robust preselection based on overlaps between element's bbox (shadow-element's bboxes in case of quadratic elements) and side's bbox.

added dserror in CollectElements() function of pointpool(octTree)-class to avoid the usage of this preselection in the future


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16913 060e22d8-9d1c-0410-868a-e3a950d63d95