o output for optimization field working now o opti discretization now also called "opti" -> adapted input files
One remark:
An element has to be in the "ParObjectList" if it shall be used in postprocessing by filters. This list is actually only a cout but otherwise the derived elements cannot be used, resulting in a very unclear error message. This error message is improved a little.

Thanks to Georg for his help ;)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@16929 060e22d8-9d1c-0410-868a-e3a950d63d95