Monolithic FSI with matching grids: =================================== - changed sign convention for Lagrange multiplier such that lambda = nodal forces onto structure for both fluid and structure split - This does not affect the solution in the primary variables displacement, velocity and pressure. Only the Lagrange multiplier is affected. - removed unused class variables - removed duplicated header inclusion in cpp-files
Please note:
In the matching grid implementation, the coupling matrices are assumed to be identity matrices due to historical reasons. When doing so, the Lagrange multiplier field cannot be interpreted physically as the traction at the interface. It just represents nodal forces. When using the Mortar framework, the Lagrange multiplier field really represents the traction at the fluid-structure interface. Nodal forces are then obtained by multiplication with Mortar matrices.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17126 060e22d8-9d1c-0410-868a-e3a950d63d95