Monolithic FSI with non-matching grids in structure split: ========================================================== - temporal consistent exchange of forces between the fields - Check, how good kinematic constraint and dynamic equilibrium are satisfied at the end of a time step - Lagrange multiplier 'lambda_' at the interface is written for output and restart - Lagrange multiplier 'lambda_' at the interface is read for restart - fixed compiler warnings - added some Teuchos:: and std:: - changed sign convention for Lagrange multiplier such that lambda = traction onto structure - added some Teuchos:: and std:: - some terms in the recovery of lambda are commented out as also done for the matching grid implementation (reason is given in comment in the code)
Please note:
In the non-matching grid implementation using Mortar, the coupling matrices M, D and P are computed by the Mortar framework. When doing so, the Lagrange multiplier field can be interpreted physically as the traction at the interface. We use the convention, that the Lagrange multiplier field equals the additional FSI traction at the fluid-structure interface onto the structure. Nodal forces are obtained by multiplication with Mortar matrices.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17157 060e22d8-9d1c-0410-868a-e3a950d63d95