towards global optimization of material parameters
this extends the statistical inverse analysis framework represented by the
StatInvAnalysis class:

* the functionality of StatInvAnalysis is reduced to handle the computation of
the gradient of an objective function w.r.t to specific material parameters.
Thereby, the handling of spatially varying (so far elementwise; later on
parametrized) material parameters is encapsulated into an extra class.
Similarly this is done for the objective function. Various materials as well as
objective functions can be added. So far this only works for the AAA_stopro
material.

* in order to be able to easily extend to further materials and also to decide via
inputfile, which parameters of which material are to be optimized, MAT::PAR::PARAMETER is
extended by a virtual method to return the valid optimization parameters (to be
implemented in the specific material). The Material class now has a virtual function
enabling the setting of element specific material parameters by name

* the eventual optimization is fowarded to specific classes derived from
StatInvAnalysis. So far this is a very basic gradient descent scheme and a dummy
class to hold a MonteCarlo sampler

* the gradient information is computed via adjoints restricting the use to
materials for which the adjoints are explicity implemented. (at least as long as
gradient information is used to find some optimum). FD gradient approximations
are not excluded to fit into this framework easily

this is still pre-alpha version, but it compiles and doesnt mess around


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17471 060e22d8-9d1c-0410-868a-e3a950d63d95