Added contructors
LINALG::SerialDenseMatrix(Epetra_SerialDenseMatrix& Source,
                          Epetra_DataAccess CV, 
                          bool set_object_label=true);

LINALG::SerialDenseVector(Epetra_SerialDenseVector& Source, 
                          Epetra_DataAccess CV);

to be able to view Epetra_SerialDenseMatrix/Vector as 
LINALG::SerialDenseMatrix/Vector. This can be done by setting the "CV=View",
identical to the use in LINALG::Matrix. 

Instead of setting "CV=Copy" for a deep copy, one should rather skip this and
call the sopy-constructor with the Epetra_SerialDenseMatrix/Vector alone,
because this can also be called with "const" objects.

Additionally modified copy-ctor to get the Label_ of Epetra_Objects consistent
with common use.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17534 060e22d8-9d1c-0410-868a-e3a950d63d95