Fixed a bug regarding the solidh8fbar element.
Problem: For Volume Neumann boundary conditions regarding the solidh8fbar element, the time curve, the on/off switches and the values worked so far, but the functions (last 6 parameters of the condition definiton in the .dat file) were disregarded. 

Solution: The code was adapted to account for functions in the Volume Neumann boundary conditions for solidh8fbar elements based on the procedure for the solidsh8 element.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17610 060e22d8-9d1c-0410-868a-e3a950d63d95