Added functionality of orthogonal projection (aka Krylov projection) for direct solver. The projection is used e.g. in fluid to fix the pressure level for a pure Dirichlet problem or the concentration level or in scatra to fix the level of some scalar quantity for a pure Neumann problem. This is done by really creating the projection matrix at apply this on the systemmatrix in a double matrix-matrix product (ouch! expensive!).
Furthermore, adapted the Krylov solver to be able to deal with non-orthogonal
kernel vectors. This is the first step to handle e.g. rotational zero-energy
modes in structure problems. However up to now, projection can only be applied
to fluid or scatra problem. To change this one would need to change the valid
parameters and the solver-setup-call in NewtonFull().

The way the Krylov projector is called is now similar in scatra and fluid. The
projector and its call is also more beautiful und understandable, mainly due
to a change in structure including reduced code copying and - hello - more
comments! Header in header inclusion are reduced to a minimum, also added some
Teuchos:: for RCPs.

Took ownership of Krylov projector files.

Test case for Krylov projection for direct solvers follows.

Thanks to Tobias for answering als the stupid solver questions.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17614 060e22d8-9d1c-0410-868a-e3a950d63d95