Further step to a more flexibel and more consistent Krylov projection:
Changed validcondition-setup to NUMDOF-framework for the dat-file. Here it is
called NUMMODES and must be for fluid and scatra the number of dofs per
node. Now, also in scatra field the input in the dat-file is only used as flag
for the constant mode of the selected scalar. Theoretically, it should now be
possible to use a Krylov projection for scatra AND fluid simultaneously.

old version to fix pressure level in a 3-D fluid:
E 1 - fluid 0.0 0.0 0.0 1.0 0.0 0.0 integration

new version to fix pressure level in a 3-D fluid:
E 1 - fluid NUMMODES 4 ONOFF 0 0 0 1 integration

old version to fix level of 2nd scalar quantity in scalar problem with 3 scalars:
E 1 - scatra 0.0 1.0 0.0 0.0 0.0 0.0 integration

new version to fix level of 2nd scalar quantity in scalar problem with 3 scalars:
E 1 - scatra NUMMODES 3 ONOFF 0 1 0 integration

Adapted all relevant dat-files that were in ./Input/ and changed all Krylov
projections in scatra, fluid, and combust, accordingly.




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17624 060e22d8-9d1c-0410-868a-e3a950d63d95