Dirichlet BCs in monolithic FSI (only lung and constraint): =========================================================== From a meshtying point of view, all monolithic FSI schemes distinguish between a master and slave side. At the interface, only the master side is allowed to carry Dirichlet boundary conditions. The slave side is NOT allowed to carry Dirichlet boundary conditions. During the setup of the simulation, it is checked, whether there are Dirichlet boundary conditons on the slave side. If this is the case, we should throw an dserror(). Currently, just a warning is displayed.
To avoid the above mentioned problem, you have to be careful already when creating the nodesets in Cubit/... . Let me explain the strategy with an example: If there is a tube filled with fluid (like in the pressure wave example or in an artery) and the structure has to be fixed at the inflow with a fluid inflow profile, define the following four nodesets:
- structure surface that has to be fixed (for a surface Dirichlet condition)
- fluid inlfow surface (for a surface Dirichlet condition)
- structure interface ring (for a line Dirichlet condition)
- fluid interface ring (for a line Dirichlet condition)

Now, put Dirichlet boundary conditions on BOTH surfaces and a NO-Dirichlet boundary condition on the SLAVE interface ring.

Examples:
1. Structure split: MASTER = FLUID, SLAVE = STRUCTURE
                    set Dirichlet boundary conditions on BOTH surfaces and a NO-Dirichlet boundary condition on the structure interface ring

1. Fluid split: MASTER = STRUCTURE, SLAVE = FLUID
                set Dirichlet boundary conditions on BOTH surfaces and a NO-Dirichlet boundary condition on the fluid interface ring

Remember: To set a NO-Dirichlet condition, just put "0"s in the ONOFF-Flag in the Dirichlet input line.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17636 060e22d8-9d1c-0410-868a-e3a950d63d95