More or less massiv commit that changes the use of the Krylov projector:
Now, the projector is a member of the time integration impl. The vectors c_
and w_ are solely hold in the projector but still updated in the time
integration if necessary. The solver now doesn't get the vectors to built the
projector, but the projector itself. With this concept, Krylov can now also be
applied to solids. A test case is added, in which an unsopported cylinder
undergoes an orthopressure and is solved with UMFPACK, thus testing two
features: Krylov with direct solvers and Krylov for solids.

!!! Note !!!
The rather dubious check in linalg_krylov_projector.cpp where the kernel of
the projection tested on the system matrix gives the null vector has vanished!
This was necessary since rotational rigid body modes in solids never give a
pure zero (due to a small-angle approximation) but are still in the null
space. So using Krylov leaves you with two responsibilities: 
- make sure the modes you provide in the dat-file are true rigid body modes
- make sure your matrix-integration is as accurate as necessary for the solver

change in drt_opt only adapts the solver call
change in drt_discret makes it possible to compute the nullspace directly

This commit is dedicated to Pope Benedict XVI. May he still have a long and
healthy live to let him find the one and only true religion: BACI.




git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17665 060e22d8-9d1c-0410-868a-e3a950d63d95