* update for MueLu_contact2 preconditioner for meshtying problems.
In contrary to contact problems we can skip the segregation factory
for meshtying problems. Due to the "symmetric" condensation we are
not able to reconstruct the segregated linear system from the master and
slave dof maps only in a reasonable way. To enforce segregated aggregates
for meshtying we need a more sophisticated hack to provide the stiffness
matrix before applying the contact to the preconditioner for building
aggregates...

* remove some debug output (that i forgot to remove on Tuesday)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17696 060e22d8-9d1c-0410-868a-e3a950d63d95