Constitutive models for 3D solids =================================
From now on, all 3D solid materials derive from a pure virtual base class 
MAT::So3Material which defines a common interface for the basic material 
functions. As a consequence, we can avoid repeated if-statements in the
element routines and get rid of the element-specific so*_mat_sel functions. 
Instead, we can call, e.g., the Evaluate function of any material directly 
from the element routine. This new framework enables us to combine every 
material with every element (except the nstets which can only support
simple materials for the time being).

The following functions MUST be implemented by all derived materials:

    UniqueParObjectId, Pack, Unpack, Evaluate  

The following functions CAN be implemented IF NECESSARY (otherwise, default
do-nothing-versions of the base class are called):
    
    Setup, Update, ResetStep, ResetAll, VisNames, VisData

The following functions CAN be implemented IF REASONABLE AND NECESSARY
(otherwise, default versions of the base class throw a dserror):

    EvaluateGEMM, StrainEnergy


I tried to port all functionalities that have been available up to
now. However, there are things that are still missing (particularly 
in the "non-standard" elements). Hence, if you are responsible for an element 
or a material (particularly the fancy ones), please check whether your
element/material fully supports the new framework. In particular,
make sure that

   - all actions are supported adequately by your element

   - all necessary functions (e.g., Setup, Update, Reset, VisNames, VisData) 
     are available in your material if necessary (i.e., in case you have some 
     Gauss point data)

There are (almost) no queries for the material type in the elements from 
now on. This means that (instead of running into a dserror because your
material type is not supported) it might happen that you will call the default
implementation of the base class (which in many cases just does nothing). 
Example: if you try to run the inverse analysis with a fancy material, you
will at some point call the Reset-function to rerun your simulation from t=0 
with modified parameters. If your fancy material does not support the Reset-
function, your history data will simply not be reset. You will probably not 
notice it directly, but your results are just wrong.


Summary of changes:
-------------------

- Changed inheritance of 3D solid materials
- Changed function arguments to base class version (e.g., changed from
  references to pointers)
- Additional data potentially needed for evaluation of special materials
  (e.g., Gauss point number, element ID, plastic strains, aneurysm stuff,
  strength) is now packed/unpacked to/from the parameter list
- Moved visualization of Gauss point data to materials (-> introduction of
  VisNames, VisData)
- Changed MAT::NUM_STRESS_3D from unsigned int to int
- Replaced element-specific definition of number of stresses by
  MAT::NUM_STRESS_3D
- Renamed element action "calc_struct_reset_discretization" to
  "calc_struct_reset_all" for the sake of clarity
- Deleted do-nothing-implementations of Reset-function in several materials
- ConstraintMixture: renamed "SetupHistory" to "ResetAll" and "Reset" to 
  "ResetStep"
- Growth: renamed "ResetGrowth" to "ResetAll"
- ViscoGenMax: renamed "Setupvisco" to "ResetAll"
- MicroMaterial: moved EAS handling to Evaluate-function, saved calculated 
  density and made it accessible from outside via new function Density
- Removed soh8_mat_sel_gemm and introduced EvaluateGEMM function instead 
  (currently only supported by NeoHooke and StVenantKirchhoff and only 
  called by Hex8)
- Removed materials that are redundant or not supported/used anymore
  (Itskov, Yeoh, AnisotropicBalzani, ArtWallRemod, BioCell, Protein, CHARMM,
  ContChainNetwork, HolzapfelCardio, HumphreyCardio, LogNeoHooke, MooneyRivlin)
- Removed input for materials that have never been ported from ccarat to baci
- Removed function soh8_shapederiv as comment in header said "to be
  removed!!!"
- Removed function sosh8_gmshplotdis as it is not called anywhere


To do:
------

Since all 3D solid elements directly derive from DRT::Element (which
implements a material of type MAT::Material), we still need to cast to 
MAT::So3Material. The next step would be the introduction of a 3D solid
element base class which directly implements a MAT::So3Material. Go ahead 
volunteers ;-)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17743 060e22d8-9d1c-0410-868a-e3a950d63d95