Fixed bug in construction of mortar interface for FSI regarding the element offset for slave mortar elements.
When the interface discretization is build, an element offset for the slave elements ensures, that the 
slave elements are inserted into the discretization AFTER the master elements.  We have to be careful in cases, where elements of the field chosen as master side contribute to the interface surface/line with more than one element facet/edge like in the "flexible beam with vortex shedding" example. In these cases, the number of finite elements attached to the interface does not always equal the number of interface mortar elements. To ensure unique IDs and to avoid trouble, the number of master mortar elements is used to determine the element offset in FSI problems.

For fluid meshtying, we do not have to be that strict on this issue, since there the 'old' implementation, where the element offset is determined by the maximum gobal element ID() of the discretization, works due to the fact that there is only one large (fluid) discretization and not two smaller seperate (structure and fluid) discretizations.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17814 060e22d8-9d1c-0410-868a-e3a950d63d95