- introduce new guard flag for Q2/2013 Trilinos - print out Trilinos version description in baci header.
The frequency of official and internal Trilinos releases is
too low to follow all the detailed changes (especially in Epetra
and EpetraExt). Printing the git SHA1 hash of the Trilinos repository
allows to determine the exact version that has been used.
This helps to give the Sandia guys more detailed feedback about
possible bugs and problems.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17898 060e22d8-9d1c-0410-868a-e3a950d63d95