This is the SelfCut-library, the new part of our Cut-library. It allows our geometrical algorithms to handle overlapping, contacting structural bodies.
The SelfCut-framework runs before the Cut-routines and manipulates the cut mesh. The cut mesh, which is composed of the surface mesh of all structural bodies, is replaced by one new, consistent cut mesh.

The SelfCut consists of four components:

(1) Collision Detection
(2) Mesh Intersection
(3) Element Selection
(4) Status Inspection

In (1), the overlapping of the cut mesh is spotted. In (2), a new surface discretization is created and the old cut mesh is deleted. In (3), the parts of the cut mesh, which would lie inside a structural body, are erased. In (4), a output of the cut mesh is generated.

Furthermore, our Cut-library features a new triangulation method: EarClippingWithHoles(). This method can handle simple polygons with holes.

Finally, our Cut-library has now the potential to treat inline points differently. Instead deleting them we can work with them.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17955 060e22d8-9d1c-0410-868a-e3a950d63d95