Introducing code testing nightly testcases
testcase check_unused_input_params-ct
=====================================

The testcase calls Baci with the option -d in order to get the current parameters we use. Then every parameter is searched with the grep command besides the drt_validparameters.cpp file. If the grep command doesn't find the value, it will be added to a fail list, which will be printed out at the end of the test.

This test is just a motivation to check if the found parameters are still useful or could be deleted.

If it is okay for the parameter to be only in drt_validparameters.cpp please add the value in the script tests/code_test/check_input_params.py to the dictionary
PARAMS_EXISTING_ONLY_IN_VALIDPARAMS (if questions arise according syntax ask me)

The same procedure can be done if the parameter is currently not used, but we want to keep it. Therefore add the value to the UNUSED_PARAMS_TO_KEEP dictionary


testcase check_unused_inpar_params-ct
=====================================

The script greps the file drt_validparameters.cpp for INPAR::* values and searches them in the rest of the code. All values which don't appear there are printed at the end of the test.

Again this is just a motivation to think of the need of their usage.

If the parameter should be excluded from the search please add the value in the script tests/code_test/check_inpar_params.py to the dictionary UNUSED_PARAMS_TO_KEEP



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18000 060e22d8-9d1c-0410-868a-e3a950d63d95