add new build configuration file for FC16 with gcc 4.8.2
not tested yet.

Meant for our experiments with gcc 4.8. Not meant for the public.
Be aware, that some parts of BACI are not compiling with gcc 4.8 unless you apply
some inofficial patches and fixes (and some hacks).

GCC 4.8 produces better and faster binaries than GCC 4.4 which allows to use more features of
our processors. Furthermore the error messages are much more helpful and convenient
(especially in context of template programming).

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18631 060e22d8-9d1c-0410-868a-e3a950d63d95