Monolithic coupling of structure and 0D fluid governed by 2-element Windkessel condition enabled. Windkessel condition is introduced as structural boundary condition onto an enclosed surface over which the flux Q can be measured. The Windkessel equation reads
C dp/dt + p / R + Q = 0 (with the compliance C, the resistance R, the flux Q (=dVol/dt) and the pressure variable p).

This equation is time-consistently solved together with the standard structure, while Q=Q(d) being dependent on the structural displacement field and the external load vector fext=fext(d,p) being a function of d (follower load) and the Windkessel pressure p.

The coupling partially makes use of Thomas Kloeppel's constraint routine for evaluating the offdiagonal coupling matrices dVol/dd and dfext/dp (actually being the same) needed for the monolithic system matrix.
The routines currently are in the folder drt_constraint, but maybe could move to a general folder drt_windkessel in the future, since the framework should enable the implementation of more complex Windkessel models as the current RC one in a straightforward manner. Also, the solving is introduced into the newtonlinuzawa routine inside the structural time integration, and not into the fullnewton, even though we don't do any Uzawa stuff here.
A more general implementation into the normal Newton, also for Thomas Kloeppel's constraints, should be desirable, especially when combining stuff with different methods i.e. contact or so...

To apply the coupling, the structural bcs

DESIGN SURF WINDKESSEL CONDITIONS
E xxx - id 1 resistance 0.0 compliance 0.0

as well as
DESIGN SURF WINDKESSEL STRUCTURE COUPLING CONDITIONS
E xxx - 1

are needed, with the latter one being the surface to which the Windkessel pressure is applied (acutally, xxx should be equal to xxx! ;)).

The routine is enabled for GenAlpha, OST and GEMM time-integration.

Thanks to Matthias for some help with time-integration and to Michael for assisting with a little parallel bug... ;-)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18703 060e22d8-9d1c-0410-868a-e3a950d63d95