- created input parameter list for semi-smooth plasticity containing all algorithmic parameters, currently:
------------------------------------------------------SEMI-SMOOTH PLASTICITY
SEMI_SMOOTH_CPL                 1.0
STABILIZATION_S                 1.0
NORMCOMBI_RESFPLASTCONSTR       And
NORMCOMBI_DISPPLASTINCR         And
TOLPLASTCONSTR                  1.e-8
TOLDELTALP                      1.e-8

- added convergence check for plastic flow increment (Delta Lp) and NCP-residual 
- adapted test cases


Remark: We now check the (i) residual of the condensed system (ii) the displacement increment (iii) the NCP-residual and (iv) the plastic flow increment. To be more stringent (and inline with the contact saddle-point formulation), one should test the residual of the equilibrium equation before performing the condensation of the plastic variables. However, as we do the condensation directly at Gauss-point level, this uncondensed residual/internal force is cumbersome to get. Still, the current implementation is an improvement compared to a convergence check only based on the condensed residual and the displacement increment. 

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18825 060e22d8-9d1c-0410-868a-e3a950d63d95