BC-free boundary condition: --------------------------- Added a new boundary condition named the 'BoundaryCondition-free Boundary Condition'. It is used for counteracting spurious velocities at curved boundaries with slip-conditions. For details see Behr M., 2003, 'On the Application of Slip Boundary Condition on Curved Boundaries' (or ask me, of course).
LocSys:
-------
A flag named 'USEUPDATEDNODEPOS' has been introduced to decide whether
original node positions or the updated node positions shall be used
when evaluating the functions under 'funct' for the rotation vector
'ROTANGLE'. The new locsys definition looks for example like this: 

E 1 - ROTANGLE 0.0 0.0 1.0 CURVE 0 0 0 FUNCT 0 0 1 USEUPDATEDNODEPOS 1

Meaning that the z-coordinates of the corresponding nodes are
calculated by the function 'FUNCT1' under consideration of the updated
node positions.

The two locsys test cases have been updated as well.

Functions:
----------
A function named 'NODENORMALSPHERE' is introduced that calculates the
(analytical) normal vectors of nodes on the surface of a sphere (or
hemisphere). Example:

FUNCT1 NODENORMALSPHERE ORIGIN 0 0 0

A function named 'ROTATIONVECTORFORNORMALSYSTEM' is added that
calculates the rotation vector (needed for LocSys conditions) for a
given node and for a given geometry. In particular, the calculated
rotation vector describes the rotational transformation of the node's
local coordinate system into the 'normal system'. The normal direction
for each node is given by a function (e.g. the function
'NODENORMALSPHERE' above can be used). Example:

FUNCT2 ROTATIONVECTORFORNORMALSYSTEM FUNCTFORGEOMETRY 1



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18911 060e22d8-9d1c-0410-868a-e3a950d63d95